WORDPLEASE/requirements.txt
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==2.0.1
django-embed-video==1.1.2
djangorestframework==3.7.7
idna==2.6
public==0.0.63
pytz==2017.3
requests==2.18.4
self==0.0.25
urllib3==1.22
